pyproject.toml
[project]
name = "navigation2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagger-io>=0.18.6",
    "pre-commit>=4.2.0",
]
